+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## [4.1.0]
+
+### Added
+
+- Support for Balance API. (@podobaas)
+
+### Fixed
+
+- Typo on Address property of SourceLocation. (@podobaas)
+
+### Changed
+
+- Updated RestSharp to 106.9
+
 ## [4.0.1]
 
 ### Fixed